stage 2: widen the legal-source mask to u32
A 16-tile path has 17 source classes and the mask was a u16. Rust masks an
over-wide shift in release rather than trapping, so bit 16 became bit 0 and the
final path position aliased onto entry-from-hand, corrupting blitz and Masters
labels. Finkel has 15 classes and was unaffected.

+++ scripts/nn_sweep.py
@@ -298,7 +298,7 @@ def main() -> None:
         groups = {
             "parent": torch.from_numpy(d_packed.astype(np.int64)).to(device),
             "roll": torch.from_numpy(d_roll.astype(np.int8)).to(device),
-            "bits": torch.from_numpy(d_mask.astype(np.int32)).to(device),
+            "bits": torch.from_numpy(d_mask.astype(np.int64)).to(device),
             "best": torch.from_numpy(d_best.astype(np.int8)).to(device),
         }
         print(f"policy training set: {len(d_packed)} decisions (the full space), "
